Allow padding of copies in pipelines to be a buffer instead of a scalar (#636)
This picks up where #633 left off and enables padding to be buffers in
the pipeline API as well.

Padding can now be computed by another stage in the pipeline, and does
not need to be a scalar value. It is now essentially the same as any
other input to any other pipeline stage.