Fix undefined expressions in crop bounds and closures (#638)
Currently, we allow crops to have undefined bounds, which means there is
no crop at that bound.

We also allow things like `select(x, expr(), y)` in crop bounds, and
that means that the crop should do nothing in the case `x` is true. This
is a headache to support well. This PR removes this, so `expr()` would
need to be a `buffer_min` expression instead.

Also, don't capture undef vars in closures