Fix crop handling for aliasing (#652)
`substitute_buffer` would replace buffer values with undef if the value
wasn't defined, which is incorrect for the way the crop visitors were
using it in aliasing.

This adds an optional parameter to `substitute_buffer` that indicates
unspecified values should be taken from a different buffer, instead of
left as undef.

At first I tried to fix this by just substituting the buffer values we
want in the crop handler. However, this doesn't work, because crop
doesn't know how many dimensions the buffer has, so it can't fully
populate the `dims` with the correct values.