Every task is a loop in the thread pool (#666)
This makes every queued task in the thread pool a `parallel_for<>`
object instead of a simple task. (Tasks are now loops with `n = 1`.)

This eliminates the need for the `cancel` mechanism, which was clunky
and high overhead. Now, we can simply see if any work remains *before*
starting to work on a task instead of after, which eliminates a lot of
unneeded worker tasks, especially in the case of nested parallel loops.