Fix not waiting for loops that were not completed by main thread but still not done (#672)
#671 has a theoretical bug, if the main thread working on a parallel for
loop does not complete the loop, `work_on_loop` returns false, but the
loop might not be done yet and we need to wait for it.

I've never seen it actually cause a test to fail, but it definitely
seems like a bug.

This does a related cleanup: by moving all of the loop completion logic
to be inside `thread_pool::run`, it eliminates a use of `wait_for` from
the thread pool public API, which makes the thread pool API more
implementable by other thread pools. There is still a use of `wait_for`
in the evaluator logic, but, one step at a time.