Add optional pattern matching in the rewriter (#673)
This PR makes it possible to optionally match something in a pattern,
with a default value if not matched. This allows reducing the
duplication in many simplifier rules, which makes the simplifier faster
and smaller.

This PR is a subset of #662, removing many of the optional rule rewrites
for a later PR.