Properly separate `thread_pool` base class from `thread_pool_impl` (#676)
It's probably best to review the two commits in this PR separately.

The first commit refactors `parallel_for<>` to have an abstract base
class, so other implementers of `thread_pool` could use a different
implementation if they so choose.

The second commit moves the `parallel_for<>` and `thread_pool_impl`
classes to a separate library, that in slinky itself is only used by
tests.