Add `async` node and generate parallel tasks with `parallelize_tasks` optimization (#682)
This PR adds an `async` node that forks the program into two tasks, and
joins after both tasks are complete.

It also adds `parallelize_tasks`, an optimization that searches `block`
for independent stmts that can be run in parallel, and parallelizes
them.

This is not enabled by default yet. `parallelize_tasks` should not
parallelize everything it finds that are independent, many stmts are
cheap and the synchronization overhead outweighs the benefits of
parallelizing them.