Add `thread_pool_impl::work_until_idle` (#701)
This helper is useful when trying to use `thread_pool_impl` to wrap an
existing thread pool. This is similar to `run_worker`, but instead of
waiting on the condition variable when there is no work to do, it just
returns. This is useful to call from a scheduled task in another thread
pool that we want to work on the `thread_pool_impl` queue instead.