Fix bug when propagating inconsistent strides from multiple dimensions (#716)
When we are aliasing a buffer, we check that there are no strides on the
target allocation, and if not, we assume we can propagate our strides to
the target allocation.

However, in the case of stencil copies, this may be insufficient:
multiple dimensions of the stencil copy may try to alias the same
dimension, with different strides.

This PR addresses this issue: when considering an alias to be
compatible, we track what the strides we want to be, and only consider
the alias incompatible if the stride needs to be set to more than one
inconsistent value.