Allow copy funcs to have a custom copy implementation (#713)
This allows one to define pipelines with copies in them that might alias
to another buffer, but if they don't, to call a user defined
implementation of copy instead of `slinky::copy`.

This provides a way to optimize special cases of copies, such as
transposes that affect dimensions with a small and/or fixed stride.