Fix aliasing of slices (#726)
This PR fixes a bug where if a copy has a slice of one of the input
dimensions, the alias ignores it.

The other change is just a minor tweak for clarity, that none of the
original buffer dimensions affect the alias (all of them were
overwritten by the alias dimensions).