Add the ability to make custom call expressions (#732)
This PR makes it possible to define custom call expressions, implemented
as an `std::function<index_t(span<const index_t>)>`.

Until now, we've been making do by making call_stmt objects that compute
whatever is required.

However, this does not allow for the result to influence bounds
expressions, because there is no way to get from a buffer value in
memory to an expression (by design).

This mechanism allows this to happen.

This is a dangerous mechanism, if the callback implementation is not a
pure function of the inputs, the pipeline is likely to misbehave.