Add `scalars` field to `call_stmt` (#740)
This PR adds `call_stmt::scalars`. Currently there is no support for
this in the wrappers that `func` defines, but for `func` users that make
custom callbacks, it enables passing a scalar through a pipeline to a
func implementation.

You might think that such scalars could simply be captured by the func
callback, but this doesn't work because the simplifier and other
processes that get applied might eliminate dependencies from the
expression as unused, or substitute variables to be different variables.
This change makes those expressions visible to slinky, so that doesn't
happen.