Move repo to slinky subfolder (#746)
I *really* would rather not do this, but it seems necessary to avoid a
lot of headaches from bazel.

IMO, bazel has a design flaw in the way it manages external repos, in
that it doesn't give them a "namespace" for lack of a better term. So we
have to give it a namespace this way. This seems like a typical
convention for most repos.