Add more explicit broadcasting support in buffers (#751)
A common use case for a slinky pipeline is to compute something with
extent 1 in a dimension, and then broadcast that when consuming it.

Currently, this is doable, but tedious, by inserting a copy that we
expect to alias as a broadcast. We also currently support another form
of broadcasting, which is implicitly assuming dimensions that are beyond
the rank as a broadcast (i.e. indexing that dimension is a no-op).

It would be nice if we had something similar for dimensions not beyond
the rank: a way to say that a dimension is "missing", which would behave
like an extent 1 dimension when being produced, and behave like a
broadcast when being consumed.

That is what this PR is working towards.