A broadcast dimension should not be considered to have a stride (#763)
This is a bit of a temporary fix.

This is another example of an issue where we really need to have a
proper "undefined" `dim_expr`, instead of trying to recognize the
dimension 0, 0, 0, 0 everywhere. In this case, because the stride is 0
(and not undefined), we think the dimension can't be aliased.