Always inline `eval_context::set` to avoid msan false positives (#765)
This is a bit of a hack: `eval_context::set` is basically a swap
operation. If one of the values being swapped is uninitialized, msan
complains, but only if the compiler doesn't inline it (msan only
complains if an uninitialized value is returned from a function). This
function is normally inlined, but if the build uses `-Oz`, it doesn't
(sometimes).